requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
beautifulsoup4==4.13.4
blessed==1.20.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
editor==1.6.6
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
inquirer==3.4.0
langdetect==1.0.9
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-jose==3.4.0
python-multipart==0.0.20
pyuseragents==1.0.5
readchar==4.2.1
requests==2.32.3
rsa==4.9.1
runs==1.2.2
safeIO==1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
translatepy==2.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
xmod==1.8.1
